requirements.txt
# WealthGuard AI - Python Dependencies

# Core Framework
fastapi[standard]==0.115.12
uvicorn[standard]==0.32.1
python-multipart==0.0.20

# Data Processing
pandas==2.2.3
numpy==2.2.1

# PDF Processing
pdfplumber==0.11.6
pypdf2==3.0.1
tabula-py==2.10.0
camelot-py[cv]==1.0.0
jpype1==1.5.2

# Machine Learning & Forecasting ⭐
prophet==1.1.6
scikit-learn==1.6.0
networkx==3.4.2

# LLM
ollama==0.4.7

# Workflow & MLOps
prefect==3.3.4
mlflow==2.21.3

# Visualization
matplotlib==3.10.1
plotly==6.0.1
streamlit==1.44.1

# Utilities
pydantic==2.11.3
python-dateutil==2.9.0.post0
pyyaml==6.0.2
requests==2.32.3
python-levenshtein==0.27.1
fuzzywuzzy==0.18.0

# Development Tools
ruff==0.11.5
pytest==8.3.4
pytest-asyncio==0.24.0

scikit-learn==1.3.2      # For anomaly detection
pytest==7.4.3            # For testing
pytest-asyncio==0.21.1   # For async tests
httpx==0.25.1            # For async HTTP requests

# RAG & Vector Search (Enhanced)
sentence-transformers==2.5.1
faiss-cpu==1.8.0
